+++ packages/cmd/src/CommandProcessor.ts
@@ -31,7 +31,7 @@ import { IInvokeCommandParms } from "./doc/parms/IInvokeCommandParms";
 import { ICommandProcessorParms } from "./doc/processor/ICommandProcessorParms";
 import { ImperativeExpect } from "../../expect";
 import { inspect, isString } from "util";
-import { TextUtils } from "../../utilities";
+import { ImperativeConfig, TextUtils } from "../../utilities";
 import * as nodePath from "path";
 import * as os from "os";
 import { ICommandHandlerRequire } from "./doc/handler/ICommandHandlerRequire";
@@ -317,7 +317,7 @@ export class CommandProcessor {
                 `${CommandProcessor.ERROR_TAG} invoke(): Cannot invoke the handler for command "${this.definition.name}". The handler is blank.`);
         }
 
-        let commandLine = this.commandLine;
+        let commandLine = ImperativeConfig.instance.commandLine || this.commandLine;
 
         // determine if the command has the user option and mask the user value
         let regEx = /--(user|u) ([^\s]+)/gi;
@@ -841,6 +841,7 @@ export class CommandProcessor {
         }
         this.log.info(`Command "${this.definition.name}" completed with success flag: "${json.success}"`);
         this.log.trace(`Command "${this.definition.name}" finished. Response object:\n${inspect(json, { depth: null })}`);
+        response.endStream();
         return json;
     }
 
@@ -924,6 +925,5 @@ export class CommandProcessor {
             response.console.error(`"${handlerPath}"`);
             response.data.setObj(handlerErr);
         }
-        response.console.endStream();
     }
 }