+++ packages/cmd/src/doc/response/api/handler/IHandlerResponseConsoleApi.ts
@@ -40,10 +40,4 @@ export interface IHandlerResponseConsoleApi {
      * @memberof IHandlerResponseConsoleApi
      */
     errorHeader(message: string, delimeter?: string): string;
-
-    /**
-     * Explicitly ends a stream.
-     * @memberof IHandlerResponseConsoleApi
-     */
-    endStream?(): void;
 }